pyproject.toml
[project]
name = "lora-captioning-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.15.0",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]
